refactor: rename db/metrics.py to db/aggregation
