+++ backend/scripts/setup.py
@@ -1,4 +1,4 @@
-from van311.db.metrics import calculate_metrics
+from van311.db.aggregation import calculate_metrics
 from van311.db.seeding import seed_database
 
 if __name__ == "__main__":
